Linphone 3.2.1 - 12 05 2016 - Fix for H264 download

+++ tester/log_collection_tester.c
@@ -20,7 +20,7 @@
 	#define _XOPEN_SOURCE 700 // To have definition of strptime, snprintf and getline
 #endif
 #include <time.h>
-#include "linphonecore.h"
+#include "linphone/core.h"
 #include "private.h"
 #include "liblinphone_tester.h"
 
@@ -84,7 +84,7 @@ static ssize_t getline(char **lineptr, size_t *n, FILE *stream) {
 	*lineptr = bufptr;
 	*n = size;
 
-	return p - bufptr - 1;
+	return (ssize_t)(p - bufptr) - 1;
 }
 #endif
